// src/app/invoice-management/models/capture-invoice-submission.model.ts
/*
 * Created on Thu Jul 18 2019
 * Created by - 1149: Aashish Kumar
 *
 * Copyright (c) 2019 Fortigo Network Logistics Pvt Ltd (4TiGO)
 */
export class CaptureInvoiceSubmission {
    invoice_number: string;
    submission_remarks: string;
    submission_date: Date;
    submitted_by: number;
    enteredOn: Date;
}